Use sockets to communicate.
Each side of the network will have one TcpSocket* other;
to send and receive data to and from the other side.

For the client, this socket is created and connectToHost,
a signal is emitted when the connection is done (onConnected).

For the server, a signal is emitted when someone is connected (onNewConnection).
When there is a new connection, the server can have access to the socket to communicate.